+++ src/commands/CommonCommands.js
@@ -1,10 +1,11 @@
+/* eslint-disable semi,@typescript-eslint/no-var-requires */
 const vscode = require('vscode')
+const axios = require('axios');
 
 const Core = require('../Core')
 const Enum = require('../Enum')
-const Meta = require('../Meta')
-
-const rp = require('request-promise')
+const Meta = require('../Meta');
+const { showError } = require('../error-handling');
 
 class CommonCommands {
 	static async register(appsProvider, _authorization, _environment) {
@@ -38,19 +39,18 @@ class CommonCommands {
 						`${_environment.baseUrl}/${Core.pathDeterminer(_environment.version, '__sdk')}${Core.pathDeterminer(_environment.version, 'app')}/${app.name}/${app.version}/${Core.pathDeterminer(_environment.version, context.apiPath)}/${context.name}` :
 						`${_environment.baseUrl}/${Core.pathDeterminer(_environment.version, '__sdk')}${Core.pathDeterminer(_environment.version, 'app')}/${Core.pathDeterminer(_environment.version, context.apiPath)}/${context.name}`
 					try {
-						await rp({
+						await axios({
 							method: 'DELETE',
 							uri: uri,
 							headers: {
 								Authorization: _authorization,
 								'x-imt-apps-sdk-version': Meta.version
 							},
-							json: true
 						})
 						appsProvider.refresh()
 					}
 					catch (err) {
-						vscode.window.showErrorMessage(err.error.message || err)
+						showError(err, 'apps-sdk.delete');
 					}
 					break
 			}